Revert "MDEV-10296 - Multi-instance table cache"
This reverts commit a2b1c58c19f44ddaac3ec4d4714ffb0b2bec5f35.

+++ sql/table_cache.h
@@ -70,7 +70,7 @@ enum enum_tdc_remove_table_type
 
 extern ulong tdc_size;
 extern ulong tc_size;
-extern uint32 tc_instances;
+extern ulong tc_instances;
 
 extern bool tdc_init(void);
 extern void tdc_start_shutdown(void);
